Import and Export to disk in dev mode and to S3 in test/prod mode

+++ scrape/src/task_data_import.py
@@ -20,35 +20,40 @@ class ImportDataTask(Runnable):
             "and topic scores are excluded."
         )
 
-    def __init__(self, s3_key: str, *args, **kwargs):
+    def __init__(self, filename: str, *args, **kwargs):
         super(ImportDataTask, self).__init__(*args, **kwargs)
-        self.s3_key = s3_key
+        self.filename = filename
 
     def run(self):
-        if not self.s3_key:
+        if not self.filename:
             self.log("S3 URL needs to be specified. Abort.")
 
-        s3_client = S3BucketClient(
-            aws_access_key=settings.AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-            endpoint_url=settings.AWS_S3_ENDPOINT_URL,
-            bucket=settings.AWS_STORAGE_BUCKET_NAME,
-        )
+        local_dir = settings.DB_EXPORT_LOCAL_DIR
+        local_filepath = f"{local_dir}/{self.filename}"
 
-        local_dir = "tmp"
-        local_filepath = f"{local_dir}/{os.path.basename(self.s3_key)}"
+        if not settings.DB_EXPORT_STORE_LOCALLY:
+            if not os.path.exists(local_dir):
+                self.log(f"Create directory \"{local_dir}\"")
+                os.makedirs(local_dir)
 
-        if not os.path.exists(local_dir):
-            self.log(f"Create directory \"{local_dir}\"")
-            os.makedirs(local_dir)
+            s3_client = S3BucketClient(
+                aws_access_key=settings.AWS_ACCESS_KEY_ID,
+                aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
+                endpoint_url=settings.AWS_S3_ENDPOINT_URL,
+                bucket=settings.AWS_STORAGE_BUCKET_NAME,
+            )
 
-        try:
-            self.log(f"Download S3 key \"{self.s3_key}\" to \"{local_filepath}\"")
-            s3_client.download_file(self.s3_key, local_filepath)
+            try:
+                s3_key = f"{settings.S3_DB_EXPORT_LOCATION}/{self.filename}"
+                self.log(f"Download S3 key \"{s3_key}\" to \"{local_filepath}\"")
+                s3_client.download_file(s3_key, local_filepath)
 
+                self.log(f"Import data from \"{local_filepath}\"")
+                DataImport.import_data(local_filepath, log=self.log)
+            finally:
+                if os.path.exists(local_filepath):
+                    self.log(f"Remove \"{local_filepath}\"")
+                    os.remove(local_filepath)
+        else:
             self.log(f"Import data from \"{local_filepath}\"")
             DataImport.import_data(local_filepath, log=self.log)
-        finally:
-            if os.path.exists(local_filepath):
-                self.log(f"Remove \"{local_filepath}\"")
-                os.remove(local_filepath)
